Make a lot of progress on the generator
Maybe I'll get to it, but I don't understand why people love Go so much.
Either way, I'm going to finis this and then maybe I'll understand. I
still need to do the parsing, which shouldn't be too bad, the file
writing, which should be fine (I think), and the wrapper stuff, which
should be fine. I think. *should* be fine.