It is not working
There are some things I know are wrong. It's not parsing the home page
correctly, which I can fix pretty easily either by having a special
function for it, or maybe by using the index and using a special
function for that. Also it takes forever. I don't know if there is
something I can do about that though. Maybe not. The only trouble is
that I don't know if it's in an infinite loop or if there's just a lot
to do. Maybe I can have a set of pages that it's already done and that
would help. But there's still a lot of stuff to mess with. And it
occurred to me recently that I still have to actually do school. So it
goes.